s/vwm_xwin_is_visible/vwm_xwin_is_mapped/
Renaming to better describe what this actually checks.

Initially I was leaning towards adding a data structure for actual visibility
checks rather than just assuming all vwm-mapped windows are visible.  But
there's a need for knowing just if windows are currently mapped as well, so
this function can stay implemented this way with the appropriate name.