overlay: short-circuit draw_heirarchy_row() if unchanged
If vmon hasn't seen any heirarchical changes, and we aren't forcing a redraw,
and the process information being shown isn't changed, don't bother
re-rendering the same thing the overlay already contains.

This can be further improved, like if only wchan changed and nothing else, only
redraw the wchan column for the relevant row.  It gets tricky quickly though,
because a new wchan could be wider than the column currently permits for
example, then we'd need to go render all the rows to the new wchan column
width...

It's simpler to just redraw it all if anything has changed,  and this stuff
doesn't generally change that frequently in practice so it's pretty effective
as-is.