key: decrement key_is_grabbed release of multi-Alt
Upon releasing all keys concluding a grab the counter gets reset so this isn't
generally observed as being a problem.

Sometimes the second Alt is released by itself, restoring the origin, and the
original grabbing single Alt persists for subsequent window management
operations.  It's in this situation when the bug manifests.  If the final Alt
release occurred with a focused window/desktop differing from the origin, on
the final Alt release an unexpected restore to the origin occurred.

Usually this goes unnoticed, because typically the lone Alt release occurs
immediately following the other one, and the second restore to origin happens
to be idempotent.