overlay: fix bug with extraordinarily long argvs
This was a known bug, there's a TODO sitting right there noting it.
The items array was sized very large so it never triggered and was
forgotten about.

Running `make tags` in the linux kernel source steps on it though,
because it constructs a massive argv.

This just adds a bounds check so no crash occurs in argv2xtext().

I don't see the point of allocating memory for this as the TODO's
suggested, since any such argv is unlikely to fit in the overlay
anyways.

Also shrunk the max from 1024 to 512, which is still quite large.