overlay: fix heirarchy tee/cap drawing bug
If walking ancestors never entered the loop, bar_y isn't set.

Set the bar_y once per row, and independent of ancestors.  This
is obviously more correct and fixes a situation where the siblings
have children but there are no ancestors (siblings under the root).

Surprisingly this isn't generally observed in vwm, but was noticed
in vmon development.