overlays: paramize vwm_overlay_render() Render op
In vwm we were always doing a transparent overlay to preserve underlying
window visibility.  With vmon this is undesirable and we just want to
copy the currently cached composited contents to the window, which is
also substantially less costly to perform.

Parameterize the operation so vwm and vmon can specify what's appropriate.