overlays: add ability to explicitly set rate
In vwm it was only necessary to relatively increase/decrease the sample
rate.

With vmon being primarily a cli tool, explicit setting of the rate is
desirable.  This commit reworks things a little de-specializing the zero
value of overlays->sampling_interval, while switching this to instead of
being an index into the intervals table simply contain the float interval
itself.  The math.h INFINITY macro becomes the new paused/zero value, and
simply gets an entry of its own in the intervals table as the lowest one.