xevent: cleanup vwm_xevent_handle_map_request()
Long overdue tidying of the map request handling.

This moves all the window classifying and placement stuff into a separate
helper, adding a call to that in vwm_win_manage_xwin(), where this always
belonged.

The map request handling now just manages windows that aren't already
managed, then lets the usual "is this window mapped?" logic filter the
map request.

This should fix a lingering bug where a window on the unfocused desktop
would become spuriously visible if the client mapped it.  Firefox started
doing this recently when a page finished loading.