clickety: always (re)focus clicked window
Currently vwm doesn't notice when something else takes the input focus.
In some situations, this can leave a desktop without any focused
windows, and no way to focus one because the only window is already
focused as far as vwm knows.

As a stop-gap solution, remove the !focused condition to the
focus-on-window-click block.  This way one can always force a window to
be focused by simply mod-clicking the window.