window: split out focused window setter
vwm_win_focus() previously simultaneously told the X server to set the
input focus on the window (if mapped) and set the internal vwm state of
the window to focused.

These are really two separate operations:
1. request the X server to focus the window
2. change vwm's concept of the currently focused window

Since clients can call XSetInputFocus() as well, there's need for doing
step 2 discretely in response to FocusIn events.

Nothing is functionally different after this commit, it just exposes
step 2 as a separate vwm_win_set_focused() function for a future commit
to leverage in handling of FocusIn events.