window: vwm_win_focus_next() skip shelf unmap/map
When the shelf isn't the currently focused context but the focus
next is relative to the focus context the mapping manipulation
shouldn't be performed.

This would cause unexpected appearance of shelved windows in a
virtual desktop should a shelved window become invisibly unmanaged.

Simple way to repro is to shelf an mplayer window.  When the movie
finishes, if you're on a desktop, out of nowhere the next shelved
window appears but you're not supposed to be looking at the shelf.

A simple visit to the shelf then back to the desktop would undo the
anomaly, but it's trivial to fix - hell there was already a FIXME/TODO
sitting there.