xwindow: introduce vwm_xwin_should_manage()
Preparatory commit for applying a heuristic to honoring
override_redirect.

The X11 specification more or less requires honoring this window
flag, but it's really a disaster to blindly do so.

This function will be used to evaluate override_redirect wherever
it's currently being directly used to determine wether a window
should be managed or not.

As-implemented it only ignores override_redirect when the window
dimensions match its screen dimensions (fullscreen windows).  In
the future this might get loosened up a bit to encompass windows
covering more than something unexpectedly large for a
tooltip/popup, like 50% of the screen, since valid
override_redirect uses should arguably be limited to small
windows.