launchers: integrated recordMyDesktop
This requires a current recordMyDesktop build supporting
--need-shortcuts which was just added for this use case.

You can omit --need-shortcuts to support older rmd versions, but
this is leaning on --need-shortcuts to prevent multiple rmd
instances from accidentally being executed concurrently.  Also
the shortcuts are expected to be how one pauses/stops the
capture, no vwm-level integration has been added to signal the
process for this purpose.

The default shortcut in rmd for stopping is ctrl+mod1+s which
doesn't collide with vwm's grabs, so everything Just Works.

You can utilize the vwm process monitors as a convenient way to
observe rmd's status in terms of encoding after stopping a
capture.  You can see the reduced Idle %age in the top row during
the encode, it'll go back to normal when completed.  Or you can
focus the vwm console (the attached screen session) and actually
observe the rmd output to watch the encoding progress - one of
the advantages of vwm's console.  Furthermore, you can also
observe the CPU use of recordMyDesktop in that window with the
monitors active.

The key bindings as of now are:

Mod1+Backspace: record the focused window (root window if there
are no windows / empty desktop is focused)

Mod1+Delete: record the whole desktop / root window, even if
there are windows focused, unlike Backspace which will capture
just the focused window if there are windows.

As-is this doesn't specify --full-shots, so it's always in
damage-tracking mode, which might not work well with OpenGL style
captures.  There needs to be a way to specify variants in
launchers.def with modifiers, then there could be a +Shift
variant for these where --full-shots is added, or something.

Ideally there should be a pop-up dialog where you get an
opportunity to manipulate the flags passed on from a set of
options, but I'm not doing that right now.  That way we could
toggle stuff like sound/no-sound, tweak the FPS rate, toggle
full-shots/damage-tracking, sound/video quality, etc.  Maybe it's
best to just use an rmd front-end for that, but it feels like
it'd be nice for vwm to just have a generic little dialog
mechanism for launchers, then launchers.def could describe the
parameterized args for the dialog to present controls for.

^^ TODO