*: switch to gplv2
Part of the reason for adding headless support in vmon is to
facilitate embedded use cases.  These are often incompatible with
anti-tivoization aspects of gplv3.

I am the copyright holder of all this stuff so it's entirely fine
to switch to gplv2.  Phil Freeman contributed one trivial patch
(4183fbd), regardless I checked if he had any objections to the
gplv2 switch and he had none.

So here we go, gplv2 all the things.