libvmon: assume short positive reads found EOF
This avoids a bunch of pread() returning 0 EOF-finding calls.

These are proc files, and actually shouldn't be getting read in a
loop like this at all because it's racy to do so.  With proc
files you need to atomically read everything you wish to parse as
one sample as an atomic unit.

So this needs to be properly reworked to enlarge the buffer when
a read exhausts it, throwing away what was read, then repeating
the sample with the enlarged buffer.

But this is tenable for now until I get around to the proper
rework... just looking to reduce some sampling overheads on lower
end embedded devices.