libvmon: tidy up some elastic buffer handling
Much of this stuff is pretty sloppy ever since it was first
written as a casual experiment.  Fixup types to use
size_t/ssize_t where appropriate, and free the actual realloc'd
member of the char_array struct - the container struct with the
assumption that the member is at the struct start.