vmon: add --wip-name commandline argument
--wip-name sets a constant name to use for the pre-rename temp
name of snapshots.

Otherwise a dynamic dot-filename augmented from the destination
name is used.

The impetus for this is to simplify garbage collection of
produced PNGs.  We'll ignore dot-files to never step on the WIP
file during GC, but would like vmon to truncate and reuse the
same WIP name so it's not potentially accumulating the dot-files
in weird failure modes where the snapshots repeatedly fail to
complete and get renamed.

So in such a scenario we'll pass a dot-file for --wip-name and
vmon will keep reusing that name, never leaking more than the
single WIP file in even a persistent a failure mode.