vcr: clear shadow layer for entire row
This should have been changed when transitioning to nibbles and
the separators became generated @ present time when serializing
to png.

The first pass clears the shadow layer while assigning the first
offset part of the shadow.  But it was staying within the
separators, which is now resulting in small bits of littered
shadow corruption in the separators barely visible in the PNGs.

Fix is trivial; initialize the separator bits when shadowing the
row as well.