libvmon: maintain a flag indicating if a process is threaded
Preparatory commit for enabling charts to apply % scaling to
non-threaded procesess, to make better use of the row's available
space.

A non-threaded process can't use more than a single core, so it
should be able to scale its %age out to the full row height.  The
same will be applied to individual thread rows, as those can at
most use a single core.

The exception is a threaded process - its CPU %ages are
aggregate, and must represent up to the number of CPUs in the
system within their row.