charts: scale % bars by num_cpus when appropriate
For rows reflecting threads and single/non-threaded processes,
let's scale the bar % by the number of cpus, so they can use the
full height of the row.

These tasks can't scale to multiple CPUs, so it's pointless to
leave vertical space for the other cores' capacity, if present.

For multi-threaded process rows, the vertical space continues to
accomodate all cores.

I've been on the fence about this change for a while because it
increases the cognitive load of reading the graphs, now the
scales are inconsistent.  But when you've got 16 cores like on my
AMD P14s thinkpad, combined with a row height of 16 pixels, you
start wishing these rows used the full height of the row for
their single-core-constrained %ages.