charts: compute a "sample duration" for current sample
This turns the time passed since the last sample taken into a
"sample duration".

Ideally this would always be 1, and up until now in the main use
case, vwm, it's been assumed to generally be 1 and drops in the
timeline treated benign/fleeting because of the live viewing.

But with the introduction of --headless and increasing use on my
servers / embedded interests, this has become more problematic.

In this commit the duration is only being maintained, but not
applied.

Subsequent commits will have to repeat the current sample in the
graphs (this_sample_duration - 1) times.