charts: compute an "adherence" value
This is an attempt to add a schedule adherence metric which a
subsequent commit will plot in a row below the top IOWait/%Idle %
row.

Ideally the adherence metric's value would always be 0, because
we're always exactly on-time with our samples.

But what tends to happen is falling behind, or rarely being
slightly ahead of schedule (particularly with the epsilon
introduction).

This metric can serve as a sort of proxy for userspace's ability
to get scheduled on time, which is a useful thing to see.