libvmon: assume stale threads in a stale process
The charts code in vwm/vmon assumes stale descendants of a stale
node, but there seems to be some theoretical potential for that
to not hold true.

Let's be more agressive about ensuring that's the case.  The
current code already does this style propagation for stale
processes, but not for threads.  It seems like it'd be weird to
have this happen, but maybe there's edge cases where you have a
parent process exit with grandchildren threads surviving until
a later sample, creating opportunity for a inconsistency.