libvmon: ignore orphaned followed children still having parents
Now that vmon is a real thing with PID1 monitoring, there's this
detail of orphans getting inherited which libvmon has so far
largely been able to ignore because vwm doesn't care about such
things, always monitoring subtrees attached to X windows.

This fixes the vmon asserts when monitoring the PID1 hierarchy
where vcr_shift_below_row_up_one() would abort @:
>·······assert(*(vcr->hierarchy_end_ptr) >= row);

It was being triggered when orphans were found by the PID1
children following while they were still children of their
exited-from-kernel's-perspective-but-not-yet-libvmon's-perspective
parent process.

There's more work to be done to improve this situation, but it's
likely going to be quite invasive.  For now this is a simple
solution that should prevent the asserts.  We just might not see
a process being orphaned for a sample while it gets ignored for
still having its vestigial parent around as is_stale=1.