libvmon: public vmon_proc_monitor() shouldn't accept a parent
The public callers of vmon_proc_monitor() shouldn't ever be
passing a non-NULL parent as that's a pretty intimate and messy
internal aspect of libvmon.  So let's remove that altogether from
the public function for monitoring a process, and turn existing
one supporting parents into a libvmon-private function.

Trivial cleanup, there's so much more needed in libvmon since
it's the epitome of an organically evolved crusty thing.