TODO: introduce TODO-vmon.txt
This is an early stab at writing down some vmon-specific TODO
items along with some cash bounties for any frends/users
interested in hacking on this stuff for some cash on the side.

I may work on these myself, anyone interested in doing these
things and expecting to get paid _must_ email me first to avoid
racing with me or others working on the same thing.  Also I am
fairly opinionated in how these things should be done, so we'll
have to discuss the details or be prepared for disappointment.