vmon: monitor vmon's PID when reaper and executor
If vmon ran a command as its child and was told to be the reaper,
we probably want to see any inherited orphans vmon becomes
responsible for reaping in the graph.

The simplest robust way to achieve this is to monitor vmon's PID
as the root instead of the pid of the command being executed.

There's an argument to be made that this is how it should always
be done when executing a command, but for now I'm only going to
do it when being the reaper.  I think there's also an argument to
be made that becoming a subreaper should also be the default when
executing a command.