window: support migrating non-autoconf windows by screen
This is better than nothing, but the screens should be cycled in
MRU order like everything else.  The way things are as-is with
the bare bones RandR integration, screens are just an array of
the underlying X type.   There isn't really a vwm-native type
representing the screen, it's just an alias for the X type.

So there should be a refactor surrounding screens and a switch to
MRU order.

Anyhow, even though it's not ideal, I've been using this daily on
a multihead setup and it's a marked usability improvement.