charts: update charts->this_sample when unpausing
There's potential for spending a whole lot of cycles filling a
huge time delta from the time things became paused to the current
time.

The output from that is pretty much pointless, so let's prevent
it from happening by simply bumping the sample time to the
current time whenever we're unpausing.