libvmon: try follow children of threads too
Long ago an assumption was made that only the main thread having
tid matching the pid would ever be the parent of children
processes.

This was an optimization done to avoid having to keep open the
/proc/$pid/task/$tid/children node for every thread being
monitored.

For the longest time this seemed to be fine, as most things
weren't forking from threaded programs, and if they were it
didn't seem to produce descendants from threads other than for
tid==pid.

But that's not the case anymore.  Running `go test` on things
it's pretty apparent that go is cloning children from any old
thread and those children are under
/proc/$pid/tasks/$tid/children, and vmon is missing them from
monitoring as a result.

The bummer here is with large numbers of threads under
monitoring, there's an added fd for every one of those threads.
Each of those fds pins 4KiB in the kernel for the seq_file
buffer, it adds up.

Note this is only the libvmon side of things, there are other
assumptions in charts.c resulting in awkward rendering of this
new children-of-threads possibility.  But other than some
weirdness in the visual details, it's surprisingly mostly correct
as-is.