libvmon: introduce VMON_FLAG_NEGLECT_THREADS flag
--hide-threads needs a way to tell libvmon to both follow threads
and neglect to actually sample them.

Note this will still allocate the vmon_proc_t for every thread,
but conserves in the open proc files department by not
opening/reading everything other than the children proc file for
threads - regardless of what other proc wants have been specified.

So front-ends still need to traverse threads of processes, and if
they want to collapse the children of threads as if they're all
children of the threads' parent, that's up to the front-end to do
so.  In the vwm/vmon case, that's a detail for charts.c to
handle, when --hide-threads is in effect.  It will have to render
the children of threads as if they're children of the threads'
parent, to hide this clusterfuck from what's visible to the user.