charts: include threads count in tree/argv column
This adds a [$n_current_threads] prefix to the argv/~threadname
part of the tree column, for non-stale processes.

It's suppressed for stale processes because we don't actually
have a meaningful count to render at that point; it'd always be
zero.  This would be visible as an oddity in the snowflakes.

The count is also suppressed for non-threaded processes, leaving
a bare "[]", which may change again in the future.  It seems like
noise in the chart, but it also serves to prevent an argv somehow
starting with "[N]" being ambiguous - perhaps not something worth
caring about *shrug*.