clickety: handle edge case where impetus is centered
This actually became likely to happen with the introduction of
the 'm' modifier for moving the pointer to the focused window's
center.

Occasionally I'll accidentally initiate a resize after a Mod1-m
to brin the mouse pointer, because I'm accessing the xterm popup
window for tweaking something like font size.  But instead of
pressing ctrl-rclick, I hit mod1-rclick, and if perfectly
centered, this would disappear the window on release.

Ultimately it's an off by one error of sorts, just give the
boundary case to the < side.