vcr: add flags for vcr_draw_text()
The current code assumes clipping text to the backend's
bounds.

I'd like to enable sticking text in the graph layers, which
really needs to support wrapping to Just Work, because of how the
graph layers are unwrapped @ present to keep the incremental
rendering of the graphs efficient.

So here let's treat the existing code as the clipped mode.
Future commits will implement the wrapped mode.